chore: ground truth for constants
