+++ src/hooks/useAnalysisController/constants.ts
@@ -1,16 +1,10 @@
-import {
-  MOVE_CLASSIFICATION_THRESHOLDS,
-  MAIA_MODELS,
-} from 'src/constants/moveClassification'
+import { MOVE_CLASSIFICATION_THRESHOLDS } from 'src/constants/analysis'
 
 export const COLORS = {
   good: ['#238b45', '#41ab5d', '#74c476', '#90D289', '#AEDFA4'],
   ok: ['#ec7014', '#feb24c', '#fed976', '#ffeda0', '#ffffcc'].reverse(),
   blunder: ['#cb181d', '#ef3b2c', '#fb6a4a', '#fc9272', '#fcbba1'].reverse(),
 }
-
-// Re-export for backward compatibility
-export { MAIA_MODELS }
 
 // Import thresholds from unified constants
 export const BLUNDER_THRESHOLD =
